move shmem test to its own file
