+++ teleprox/__init__.py
@@ -10,3 +10,4 @@ from .qt_server import QtRPCServer
 from .proxy import ObjectProxy
 from .process import start_process, DaemonProcess, ChildProcess
 from .processspawner import ProcessSpawner  # for backward compatibility (use start_process instead)
+from .shmem import SharedNDArray
