+++ teleprox/tests/test_spawner.py
@@ -1,15 +1,13 @@
 # -*- coding: utf-8 -*-
 # Copyright (c) 2016, French National Center for Scientific Research (CNRS)
 # Distributed under the (new) BSD License. See LICENSE for more info.
-import logging
 
-import numpy as np
+import os
+
 import pytest
 
 from teleprox import start_process
-import os
 from teleprox.tests.check_qt import requires_qt
-from teleprox.shmem import SharedNDArray
 
 
 def test_spawner():
@@ -41,18 +39,6 @@ def test_serverless_client():
     rmt_os = cli._import('os')
     rmt_os.getpid()
     proc.stop()
-
-
-def test_shared_ndarray():
-    proc = start_process('test_shared_ndarray', start_local_server=False)
-    cli = proc.client
-    shared = SharedNDArray.copy(np.array([1, 2, 3]))
-    rmt_shared = cli.transfer(shared)
-    assert rmt_shared.data.shape == shared.data.shape
-
-    # test closing nicely
-    proc.stop()
-    assert shared.data[0] == 1
 
 
 @requires_qt
